Fixed possible headers/libraries mismatch in extconf.rb.
When Ruby is installed in a non-system location (like a MacPorts installed Ruby on OS X) the extconf.rb script was picking the non-system libraries but was using the system headers. This mismatch was causing one of the test cases (test_scan_boingboing) to fail.

This patch ensures that libraries and headers are searched from the same root prefix.