fix for assertion failure - see http://nokogiri.lighthouseapp.com/projects/19607-nokogiri/tickets/25
it's apparently OK for libxml2 to reset the xpath context node.
mostly this doesn't affect us, unless '//' is encountered in the xpath query, in which case the node is set to NULL.
so, this fix works around this by using the context document directly instead of using the context node's document.