Fix dtd validation. This resolved errors of test_validate methods in xml/test_dtd.rb and xml/test_document.rb.
This fix was really hard, and evatually I decided to add a rule for dtd validation and changed test_validate methods. To validate by dtd, dtdvalid option must be specified. Since dtd validation is done only when a document is parsed, nokogiri needs to know whether the document will be validated later or not when read/input the document. Perhaps, adding dtdvalid option is the best answer.