Changes test assertion of html/test_document#test_title=() in case of pure Java version.
To parse HTML document, pure Java version uses NekoHTML, which balances tags by default. While balancing tags, <head></head> tag is inserted if it is missing. This behavior could be turned off, but that breaks others. Thus, we'd better keep balncing tags behavior on for HTML document. This is the reason that the assertion was changed.