Prevent buffer overflow in xml_io.c when io return string is too long
If an IO object returns a string that is longer than expected, io_read_callback() writes past the limit of the buffer and crashes the interpreter. This patch checks against the size of our buffer before copying the data over.

I can't really think of a case where an IO object would do something this silly, but seems like a good practice anyways.