Fix a memory leak when XSLT.parse fails
In xslt_stylesheet.c, the method parse_stylesheet_doc() gives libxslt an error
callback and then invokes xsltParseStylesheetDoc.

Nokogiri's error callback invokes rb_exc_raise() which prevents the function
from returning to the caller. As a result libxslt never cleans up after itself
and we leak memory.

Additionally, xsltParseStylesheetDoc requires that the caller frees the XML
document on failure.

This commit fixes both of the above issues. Note one side effect -- we no longer
provide the exact reason for the parser failure in the exception. I added
rb_warning() to the error callback so that the error can be seen when Ruby is
running with $VERBOSE.

I wrote up description of how I found and fixed this leak at:
http://holymonkey.com/how-to-find-and-fix-a-memory-leak-in-a-ruby-c-extension.html