Plug memory leak when an xpath returns namespaces
When evaluating an xpath, we get a set of nodes and, potentially,
namespaces. The nodes are pointers to document nodes which will be
freed automatically at document deallocation.

The namespaces, however, belong to the xpath evaluation and it is
our responsibility to free them. At deallocation time we don't
know whether the document has been freed already, and it is unsafe
to traverse the node list to free our namespaces.

This fixes the issue by creating a separate list of namespaces and
freeing them at deallocation time.

An attempt at plugging this leak was made with commit dbda52b and
reverted with commit eacd089