Don't free namespace nodes when removed from a node set.
We free all namespace nodes at GC of the node set.

This prevents a potential segv when we have a reference to the
removed namespace node. A unit test for this has been added to
test/xml/test_node_set.rb.

Another potential segv is during GC, and was introduced by
2675a7527c. I was unable to create a reasonable unit test for
this one. This script triggers it:

loop do
  xml = Nokogiri.XML('<foo xmlns:n0="http://example.com" />')
  list = xml.xpath('//namespace::*')
  list.delete(list.first)
end