Don't segfault when a parsing error leaves a broken node reference.
Workaround for a libxml2 bug where a parsing error may leave a broken
node reference in node->doc->children.

This workaround is limited to when a parse error occurs, the document
went from having no children to having children, and the context node is
part of a document fragment.

A patch has been submitted upstream as well:
https://bugzilla.gnome.org/show_bug.cgi?id=668155

This addresses issue #600 reported by @benmiles. Thank you for the
wonderful bug report!