Narrowing conditions for which we defensively dup-unparent-and-root text nodes in the expectation they may be merged. Closes #595.
Note that this simply omits the case where we are inserting a previous
text sibling, and our case was duping and unparenting the *next* text
sibling.

It's not totally clear to me, based on looking at the libxml2 code,
that this is necessary if the pivot node is not a text node. That is,
we may be being *too* defensive here.

We should probably schedule some work to rethink all the kludges we
have around node reparenting, because I'm pretty sure we can simplify
(or decompose) `reparent_node_with` with a little thought.