Fix read-after-free bug when deallocating node sets.
This bug was reported by valgrind.

xmlXPathNodeSetFreeNs() traverses nodes in order to check
whether a namespace needs to be freed when deallocating a node
set. However, we cannot guarantee that the nodes are still
there at deallocation time.

This commit moves the tests done by xmlXPathNodeSetFreeNs into
Nokogiri_wrap_xml_node_set, at which point it is safe to do so.