Test for fragment node before assuming a node doc pointer is a doc.
The doc pointer on an xmlNode is not always an xmlDoc. The child of a document
fragment will point to the document fragment, which does not behave exactly
like an xmlDoc and doesn't have a nokogiriTuple.

This changeset tests for XML_DOCUMENT_FRAG_NODE before assuming that the doc
pointer is an xmlDoc.

This changes the NOKOGIRI_ROOT_NODE and NOKOGIRI_ROOT_NSDEF macros to functions
which test for this case.

Also check for this case in Nokogiri_wrap_xml_node.

This changeset adds a unit test which was copied from Kristi's bug report.

Reported-by: Kristi <kristi.dev@gmail.com>