Fix the tests that I broke in jruby 1.9 mode:
Wrap the EncodingReader in a java wrapper to handle the case when encoding detection succeed
and the stream is missing the first 1K of data (in which case we cannot use the underlying stream
directly, and instead rely on the EncodingReader::read() to do the right thing). the other case
is when the encoding detection doesn't succeed and EncodingReader doesn't store the first chunk
of data. In this case the wrapper will store the first chunk of data and return it the next time
NokogitiEncodingReaderWrapper::read() is called.