XSLT transform failure detection now green for Java as well
I had to mangle the test a little bit, because it looks like Java is capable of detecting the errors at instantiation, but libxslt detects them at first-use. So the block in the test needed to be expanded to cover both cases.

This commit also fixes a bug in the Java code where a ruby exception was being created but not thrown when the XSLT document creation failed.

In order to provide the user with actual useful feedback in Java, the NokogiriXsltErrorListener will need to be modified to record more than the last error message.