Fix issue #771.
The Java and C extensions are showing discrepancies when formatting xml: when using the `noblank` option, text nodes are indented and put on a newline by the Java extension, whereas the C one favours a compact approach with no whitespace nor indentation around the text node.  This commit forces the Java extension to behave like the C extension by *never* indenting and “newlining” the text nodes, whether `noblank` is used or not — in line with the observed C extension behaviour.