fix #835. replace the fix for #719 from 0d8c3f2 with a simpler fix (more info below).
The fix for #719 made ReferenceEntity nodes look like text nodes. This
caused normalization of the document (i.e. concatenating adjacent
nodes) to combine the reference entity node with any adjacent text
nodes and create a new text node. When converting the document to
a string it was impossible to know if the text came from a reference
entity or text node, or both, which caused entities to be printed
without the delimiters '&' and ';'.