Remove libxslt patches that modify configure.in.
Modifications to configure.in trigger a re-run of autotools,
which adds additional dependencies to nokogiri. This breaks the build
of libxslt when there is a version mismatch between the one used to
release the libxslt tar file and the installed one.

The removed patches itself are IMHO not necessary for nokogiri.