Cleanup build tasks for cross compiled binary Windows gems.
Remove cross ruby version checks, since they are obsoleted with the
verify_dll method.

Add verify_dll to the standard rake-compiler 'cross native gem' tasks,
so that it's more compatible with the package_win32_fat_binary command of
the rake-compiler-dev-box.

Suggest using the rake-compiler-dev-box if no cross rubies are installed,
because this is probably the most easiest way to build a proper cross
compilation environment.