(JRuby) Clear the XPath cache on attr removal.
Note we eliminate the JRuby native implementation of
`#remove_attributes` and instead make a private method to clear the
xpath context, which is called from `node.rb` if we're running on JRuby.

I've also added some test coverage for the return value of Attr#unlink
and Node#remove_attribute to make sure we get the attribute back, and
it's not being GCed out from under us.

Fixes #1109.