Use Pry rather than IRB if it's available
Pry offers a couple of features, namely the pretty-printing of
collections and syntax highlighting of returned values, that make
working with Nokogiri documents in the REPL more user-friendly.

This pull request doesn't add Pry as a dependency of Nokogiri or change
the default behaviour (I figured that was too far), but instead simply
uses Pry over IRB if Pry is available and otherwise falls back to the
previous IRB behaviour.