Do not use strdup() in Node::get() anymore.
This patch replaces the use of strdup() in xml_node.c with a proper
strncpy() solution. Besides, it adds string type checking before the
string access on 'rattribute'.

strdup() is not available in C99, hence the implicit declaration
warning. On older systems (e.g., on CentOS 6 with GCC 4.4.7, glibc 2.12)
it has sometimes failed to allocate the new string for me, causing the
subsequent access with strchr() to raise a SIGSEGV.

This crash has existed since Nokogiri 1.6.2-rc1 and up to 1.6.6.2.