Fix native build on Windows.
This depends on a fixed version of mini_portile in regards to the handling
of configure_options. I therefore assigned a git link in the Gemfile, temporary.

Rubys stdlib shellwords is generelly an issue on Windows. It's sadly, but Windows
uses several different ways to escape command line options. While shellwords mostly
works in Msys bash (and therefore in a Makefile), it is completely wrong in cmd.exe
or msvcrt based programs. It's annoying, but escape rules in mkmf.rb functions
(like try_compile etc) are different to that of $CPPFLAGS etc. which run in a
Makefile. I made an attempt to codify the different escape rules in:
   https://github.com/larskanis/shellwords/tree/master/lib/shellwords

As a compromise I used String#inspect for the lists of patches. This works well
enough on all platforms.