fixes xsl transformation in JRuby
This is one of the failures that resulted from #1212. As it turns out,
before the change the `test_transform_with_output_style' passed
trivially since the returned string was empty. This commit include a new
assertion to make sure we don't return an empty string.

This commit also change the regex in `test_transform' to match `<h1/>'
since this is the way xalan serializes the transformed document.