Revert capturing of errors while xmlCopyDoc()/xmlDocCopyNode() and sync behavior with JRuby.
This was introduced due to https://github.com/sparklemotion/nokogiri/issues/1196
and https://github.com/sparklemotion/nokogiri/issues/1208 .

However it turned out, that the change in libxml-2.9.2 was a regression, that was
fixed in: https://bugzilla.gnome.org/show_bug.cgi?id=737840 and libxml-2.9.3.

If I read the libxml sources right, it seems, that xmlDocCopyNode() is not
intended to emit any such warnings at all. Only errors leading to a failure
of the function are emitted. However these errors should be reported to ruby
space in the form of exceptions (this is not yet implemented - currently either
nil is returned or a generic error text is raised).

This patch also synchronizes the behavior on MRI to that of JRuby, so that
the error list is filled from the parser only and that it is shared after
Document#dup .