Fix encoding tests and move HTML tests to html file.
Encoding of the various text returning functions was tested
against the document encoding, but it should be UTF-8 per
definition.

The XML/HTML documents are changed to a non-UTF-8 versions,
to make the tests more meaningful.

This also fixes a test failure on Windows, that was caused
by loading a HTML document per File.read(). This function
returns a string with some cpXXX encoding on Windows, which
was used as the document encoding later on, so that the test
failed.