download tar files in parallel
We can download the tar files in parallel so that when it comes time to
compile libxslt, we don't have to wait for a download.  This should help
out people who have slow internet connections.

Down side is that the output looks weird because the progress bar is
being overwritten.