Fix #1436 - compatibity with gem-compiler.
The extconf.rb clean task removes most of the files, that are
used only while build. However this broke gem-compiler, because
it removed the archives as well, although they are part of the
gemspec.

With this patch the archives are kept. This however increases
the disk space used by nokogiri from 21MB to 30MB after install.