fixes `adopt_as' to relink the namespace of the new node
previously `adopt_as' would call `relink_namespace' on the parent
node. this caused a loop such as the following

  1000.times {|i| root << root.create_element('randomNode') }

to run in O(n^2) instead of in linear time, since relink_namespace was
relinking the root node + 1 child node initially, then the root node + 2
nodes when the second node is added, etc.

fixes #692