Look for bundled libs in lib64 else lib
This adds awareness of the lib64 build convention. On openSUSE, MacOS, and perhaps other distros, the bundled libxml2 and libxslt libraries get built in a lib64 subdirectory per distro convention.  This commit changes Nokogiri's hardcoded 'lib' convention to first check for bundled library existence in lib64, and if not found, falls back to lib.