Work around a bug in gumbo-parser's error printing.
gumbo tries to add the line that contains the error to the error
message. If the error starts with a newline, for example, if the input
is the string "<\n", then due to a bug, gumbo produces a `(ptrdiff_t)-1`
size string "starting" at the byte after the newline.

This patch modifies the `GumboError` to point one byte earlier if it
points at a newline (that isn't the first character of the input).