Download zlib tar from 'fossils' directory.
The usual download path contains only the latest version of zlib
and older versions are available at the 'fossils' path only.
If a new version of zlib arrives, all current nokogiri source gems
are immediatelly no longer installable, when running on Windows.
Fetching the zlib tar from the fossils directory solves this.

The latest zlib version is available in both directories.

The downside is, that we need to have an eye on zlib updates,
because appveyor will no longer fail when newer versions were released.
However this is the same as with the other bundled tar files.