Appveyor: Add ruby-2.4 and ruby-head to build matrix
Testing both x86 and x64 builds of all ruby versions isn't very useful,
so let's alternate between x86 and x64, to decrease test time.

Ruby-2.5 is not yet available on Appveyor, but ruby-head should be
enough for now.

Bundler is pre-installed on Appveyor, but missing on ruby-head.

Thanks to MSP-Greg for initial patch!

Closes #1702