Node#dup can now copy directly to another document
This is an optimization of what's usually done, which is to make a
document-local duplicate via `#dup` followed by an `#add_child` or
similar. The drawback in that scheme, of course, is that unparented
nodes cannot be GCed and so occupy memory while the document is still
accessible.

Using this new variation of `#dup` we can create a duplicate in the
other document directly, without a direct clone.