Handle ASCII-8BIT encoding on fragment input
If we don't know the encoding, we default to using UTF-8.  I think this
was correct behavior on 1.8, but now that Ruby is encoding aware, I
think we should raise an exception if someone passes a document encoded
as ASCII-8BIT.  However, to remain backwards compatible, we'll just
assume UTF-8 for now.

Fixes #553