add test demonstrating libxml2 XSS vulnerability
libxml2 >= 2.9.2 fails to escape comments within some attributes. It
wants to ensure these comments can be treated as "server-side includes",
but as a result fails to ensure that serialization is well-formed,
resulting in an opportunity for XSS injection of code into a final
re-parsed document (presumably in a browser).

the offending commit is:

  https://github.com/GNOME/libxml2/commit/960f0e2

we'll test this by parsing the HTML, serializing it, then
re-parsing it to ensure there isn't any ambiguity in the output
that might allow code injection into a browser consuming
"sanitized" output.

complaints have been made upstream about this behavior, notably at

  https://bugzilla.gnome.org/show_bug.cgi?id=769760

and multiple CVEs have been declared and fixed in downstream
libraries as a result, a list is being kept up to date here:

  https://github.com/flavorjones/loofah/issues/144