Handle U+0000 NULL without crashing
Gumbo's error printing code had a bug that incorrectly asserted that
characters would be nonzero.

Separately, the HTML spec specifies that a parse error occurs when a
U+0000 NULL appears in a bogus comment which Gumbo didn't didn't do.
This fixes that oversight.

Finally, the error messages are improved when handling NULL bytes.
There's still a fair bit of work to do on producing reasonable error
messages, but I think it's reasonable to say that error message values
aren't part of our API so we can change them later.