Add flags to control how nokogumbo is built
There are three configurations
1. --with-libxml2 forces nokogumbo to be built with the libxml2 headers;
2. --without-libxml2 forces nokogumbo to be built without the libxml2
   headers; and
3. (default) try hard to build with headers, but allow that process to
   fail and build without the headers.

By default, Nokogiri packages libxml2 as part of the binary. Current
versions of Nokogiri delete the libxml2 headers as part of the install.
This first checks to see if they haven't been deleted and if not, it
uses them. If they have been deleted, it downloads, configures, and then
installs the headers (in the `ext/nokogiri/.../libxml2` directory) of
the corresponding version.

However, if Nokogiri hasn't packaged libxml2, then it used the system
libraries. In this case, use `pkg-config` to locates the library/headers.

Hopefully, Nokogiri will start including the headers again at which
point the `download_headers` function can be removed.