ensure that, after reparenting, we re-use existing ns defs
previously, it was possible for a node's `ns` pointer to be
referencing an nsDef that had been removed after reparenting an orphan
node with a namespace (that removed nsDef would be rooted in the doc's
`unlinkedNodes` hash, and not referenced from the doc).

this is unrelated to the ns memory leak addressed in the previous
commits, but is something I spotted along the way.