Fix tests on JRuby 9.2
This appears a result of commit
jruby/jruby#8e631ca1d2f302a642d28b8b96427cd5e29ea222 which introduced
RubyException.toJava.  It overrides the behavior of RubyBasicObject.toJava and
returns the null throwable instead of returning the RubyException itself.

I don't understand the reason for the change in JRuby or why we used get()
instead of entry().  I think `entry' is the right method to call since it just
return the RubyObject.

[fixes #1818]