concourse: docker images need the latest gems
make sure that an update to Gemfile.lock will create a new layer in
the docker image.

also make sure all docker images are using the latest bundler.

this setup means the dev system generating the docker images needs to
be on the latest bundler and has run `bundle update` recently.