Add initial RuboCop configuration files
Working with RuboCop configuration files allows to define the
configurable attributes of each cop, and to track the files
that still don't pass the RuboCop cops.

Add RuboCop configuration files, generated with:

bundle exec rubocop --auto-gen-config

Exclude subdirectories 'tmp' and 'vendor', not used to store ruby
code of the project.