Centralize error handling
`gumbo_lex` and `handle_token` return false to indicate an error was
discovered while handling that particular character or token. This is
brittle, poorly tested, easy to get wrong, and confusing to read.

Since all errors get added via `gumbo_add_error` which has access to the
`GumboParser`, this commit adds a member to `GumboOutput` which clients
can use to determine if the document had an error, even if the
`max_errors` option is set to 0 (and thus the `errors` vector isn't
populated).