Bind the verify task for cross nokogiri.so to stage task
The previous task was also triggered in "rake compile" on the native platform.
Since the library dependencies are different then, compared to the cross build, it failed in this case.
Binding to a task that is only invoked in cross builds solves this issue.

It is the better choice in any case, since the new task is triggered at every cross build, even on re-runs.