Fix content-sniffing for Unicode strings
"\xEF\xBB\xBF" is a valid UTF-8 string of length 1. It will never
compare equal to the first three bytes of the input.

```
irb(main):058:0> "\xEF\xBB\xBF".force_encoding(Encoding::ASCII_8BIT) == "\xEF\xBB\xBF"
=> false
```

Instead, use `string#bytes` to examine the first few bytes of the
input to decide if the unicode BOM is present and to set the encoding
appropriately.

Additionally test that we're not modifying any strings that the user has
passed to us since we're using `string#force_encoding` in a few places.