Fix binary compat with musl based Linux
The binary nokogiri gem built with rake-compiler-dock-1.0.0 segfaults on x86_64-linux-musl.
This happens in:
  lib/nokogiri/xml/reader.rb:92:in `namespaces'
It turned out that any calls to sprintf() with %-arguments fail.

I did not dig deeper on assembly level, since this is the only use of sprintf in nokogiri.
Moreover it can be replaced by calls to ruby string functions easily.

Related to #1983