Fix failing test case on musl based Linux
Musl doesn't recognize cp932 encoding.
It also fails when used on the command line with iconv.
For this test case cp932 can be replaced by Shift_JIS, which is identical for the characters in question and that is handled by musl.

Error was:

  1) Error:
Nokogiri::HTML::TestDocumentEncoding#test_encoding_without_charset:
ArgumentError: invalid byte sequence in UTF-8
    /home/lars/comcard/nokogiri/test/html/test_document_encoding.rb:26:in `test_encoding_without_charset'

Related to #1983