perf: Always compile libxml2 and libxslt with '-O2 -g'
This is the default for those libraries, but we were overriding CFLAGS
in some cases.

Additionally, we need to compile with -U_FORTIFY_SOURCE to avoid
Ubuntu's convention of setting -D_FORTIFY_SOURCE=2 when -O2 is
set. This leads to problems when running precompiled libraries on
musl (see #2100 for details).

Closes #2022.