fix: some Node reparent methods parse in context of parent
The choice of context node has no functional impact for Nokogiri per
se; but it's necessary for Nokogumbo, which has a better HTML parser,
to ensure the parser is in the right state.

When testing this change against nodes without a parent, the test
suite found memory problems related to libxml2 manipulating pointers
on pickled nodes. As a result, the impacted methods now raise a
RuntimeError if they're called on a node without a parent (the concept
of a sibling or a replacement/swap doesn't make sense without the
context of a parent to hold it together, anyway).

Note that `Node#coerce` is now a __protected__ method (previously it was __private__).

Impacted Node methods:

- add_next_sibling, previous=, and before
- add_previous_sibling, next=, and after
- replace and swap

Fixes https://github.com/rubys/nokogumbo/issues/160