feat: Introduce variations of CSS::XPathVisitor that use builtins
This allows us to continue to generate XPath that uses standard XPath
functions, but also allow us to inject the preference to use the
optimized builtin implementation.

This also leaves open a path to extracting Nokogiri's CSS parser into
a separate gem, and allowing Nokogiri to inject builtin via a custom
XPathVisitor class.