doc: improve Document::NCNAME doc string
even though this constant should _probably_ be private.

Also, be explicit that the "." is not a metacharacter by escaping
it (even though its presence in a character class doesn't need to be
escaped), and add test coverage to clarify intent.

Finally, add an appropriate capture group to NCNAME_RE so that we
don't have to do an additional split() to retrieve the ncname.