Remove the default allocator function from Document
We expect that XML documents will be of type T_DATA as they are backed
by a C extension.  It was possible for people to call `allocate` on
Document and end up with a T_OBJECT.  This commit removes the default
allocator so the only way to get a document object is via the
constructors we've defined.