Hopefully fix Ruby's GC marking stack objects after return
`rb_ensure()` was being used to handle the situation that an exception
was raised during the construction of the tree. For some reason, Ruby's
GC was keeping a reference to the `ParseArgs` even after `rb_ensure()`
returned. This was causing `parse_args_mark()` to be called.

This changes this situation by
- Replacing the stack-allocated object with a heap-allocated object;
- Moving the deallocation code into the wrapped object's `free`
  function; and
- Changing from the deprecated `Data_Wrap_Struct` to
  `TypedData_Wrap_Struct`.

This change simplifies the control flow since the `parse()` and
`fragment()` functions are no longer split into three pieces.

It's likely that the runtime cost of an extra heap allocation is offset
by not needing `rb_ensure()`, but I didn't bother measuring this because
one additional allocation is trivial since there's at least one
allocation per DOM node anyway.

Fixes #2276