Remove outdated comments
The first comment is incorrect. The fragment case was changed to match
the nonfragment case in https://github.com/whatwg/html/pull/6399.

This introduced a bug that was noted by the second comment indicating
that the parser diverged from the standard to fix the bug. The standard
was subsequently fixed in https://github.com/whatwg/html/pull/6455,
obviating the comment.