meta: Github Sponsors link
Note that any money collected through this Github Sponsors program
(or through Tidelift) will be added to the Nokogiri project fund that
I track in a spreadsheet shared with some of the other
maintainers. This project fund is intended to be used, in order:

- defray project expenses (historically this is CI/CD costs)
- periodically send small gifts to contributors and co-maintainers
- give back to the open-source community

I'm open to suggestions on how to give back to the open-source
community! Please reach out to me directly if you'd like to discuss
any questions or concerns you have about funding or spending, or
transparency around this money.

[skip ci]