Correctly handle text in form elements
In the usual course of events, a `</form>` tag does _not_ pop all of the
nodes off the stack of open elements until a `form` element is popped.
Instead, it only removes the `form` element itself. Popping in the
normal fashion will flush any text node that is in the process of being
constructed.

This commit adds that flushing when `</form>` is encountered as well.

Fixes: #2450