Hoist loads and sink store in hot loop
This is from @sebpop's [PR](https://github.com/google/gumbo-parser/pull/403/).

Since `char *` can alias with everything, the compiler is unable to
prove that `output->data` doesn't alias with `output->length` and
`output` itself which forces a load of `output->data` and
`output->length` and a store to `output->length` in each iteration of
the loop. By hoisting the two loads and sinking the store out of the
loop, we're left with a loop containing just a single store to
`data[length]`.