Fix crash when calling fragment on annotation-xml elements
The HTML fragment parsing algorithm needs to know if the context
element is an HTML integration point. A MathML `annotation-xml` element
with an `encoding` attribute with a value of `text/html` or
`application/xhtml+xml` is an HTML integration point. Any other
encoding value is not. So for context elements, we check if it's an
`annotation-xml` element and if so, use its `encoding` attribute.

This commit adds tests for these two cases.

The number of parameters in the call to `[]` in order to get the
`encoding` attribute was missing which caused a crash.

Stand alone reproducer.

```ruby
require 'nokogiri'

doc = Nokogiri::HTML5.parse('<!DOCTYPE html><math><annotation-xml encoding="text/html" /></math>')
a_xml = doc.xpath('//math:annotation-xml')[0]
frag = a_xml.fragment('<p>')
```