Move rdoc and psych into their own bundler group
Psych 5 no longer bundles libyaml which causes `bundle install` to fail
on macOS which doesn't have libyaml and doesn't have an easy, standard
way to install it.

One can use `bundle config set --local without rdoc` to disable
installing the `psych` and `rdoc` (which depends on `psych`) gems.

Hopefully, this is only temporary and the upstream Psych team will
produce a better way to handle this.