ext: migrate NodeSet to the TypedData API
See #2808

Note that I chose to make a new public function,
`noko_xml_node_set_unwrap`, to allow other files to unwrap node
sets. This differs from the tactic adopted in #2807 / cb1557d0 which
was to make public the data type.

We should probably decide which is the better approach and
standardize on it.