Remove darwin linker hack
See comment in body for details on why and how it can be replaced.

The python community has been running into the same issues because
cpython extensions operate very similarly to ruby.  They got an
explanation from Apple on some of how the changes work which can be
found here and gives some additional useful context for understanding
this change:
https://github.com/python/cpython/issues/97524#issuecomment-1458855301