dev: use minitest-parallel_fork to speed up the suite
With NCPU=7 I can usually get the suite done in under 1.8 seconds on
my laptop (inspired by tldr)

Note that we stop using minitest-reporters in the main suite because
the two gems aren't compatible. I wasn't often using
minitest-reporters features, anyhow.

However, we're keeping minitest-reporters in the Gemfile because we
use it in some of the CI scripts like test-gem-installation where the
spec output is useful.