fix: turn off libxml2's new thread local storage feature
Using thread-local variables in position-independent code (i.e.,
combining `__thread` variables with the `-fPIC` compiler option)
introduces a (non-weak) reference to `__tls_get_addr` which resolves
to `ld-linux` on glibc systems and `libwinpthread` on windows systems.

This is an issue when precompiling for musl on glibc systems. If and
when we can build a separate musl native gem, we can revisit this
decision.