Treat reaching the maximum parse depth as EOF
The parser is fragile and hitting the maximum tree depth has exposed
several bugs. By treating this condition as if it is an EOF means that
we're back in the normal parsing condition.

This does expose a weakness in Gumbo's design: A spec bug that results
in skipped steps or failure to reprocess a token (or if we fail to
implement the spec in such a way), then the result can be crashes. Not
ideal.

Fixes: oss-fuzz-66107