extract CSS parser cache into CSS::Parser::Cached
The global `Parser.set_cache` and `Parser.cache_on?` methods have been
removed, as has the `Parser.without_cache(&blk)` method.

The cache is now used by `CSS.xpath_for` (optionally, via a `cache:`
kwarg) instead of being part of the parser.

CSS::Parser has a different API which accepts the selector as a
parameter to the constructor (and `#parse` no longer takes an
argument). This is a prefactor for work to come.