Fix vsnprintf usage
Standards-compliant vsnprintf implementations return the number of
characters they would have written if the buffer was large enough, not
including the null terminator. Some older versions of Visual Studio
return -1 when the buffer is not large enough. For those versions, we
need to call vsnprintf a second time with a count of 0 to get the
number of characters that would have been written. In both the C99 and
old Visual Studio case, we need to
1. Ensure the buffer is large enough to hold the string, including the
   null terminator; and
2. Increase the buffer capacity sufficiently to make repeated calls
   to print_message take linear rather than quadratic time.