css: introduce an An+B parser
which the Selectors parser will call when it consumes a Function node.

Note that this commit also extracts a BaseParser class with all of the
useful helpers and basic functionality that both parsers will need.

Finally, introduce `ensure_no_more_tokens` which can be used to make
sure we completely parse the input.